Move elixir's test to a category and divide to smaller cases
