+++ Units/parser-elixir.r/elixir-functions.d/expected.tags
@@ -1,0 +1,8 @@
+FunctionModule	input.ex	/^defmodule FunctionModule do$/;"	module
+one_liner_func	input.ex	/^  def one_liner_func, do: :baz$/;"	function
+func_no_params	input.ex	/^  def func_no_params do$/;"	function
+func_head	input.ex	/^  def func_head(string1, string2 \\\\ nil, separator \\\\ " ")$/;"	function
+func_one_arity	input.ex	/^  def func_one_arity(string1, nil, _separator) do$/;"	function
+normal_func	input.ex	/^  def normal_func(string1, string2, separator) do$/;"	function
+private_func	input.ex	/^  defp private_func(a), do: a <> " alone"$/;"	function
+private_func_no_params	input.ex	/^  defp private_func_no_params do$/;"	function
