+++ Units/parser-elixir.r/elixir-macros.d/expected.tags
@@ -1,0 +1,5 @@
+MacroModule	input.ex	/^defmodule MacroModule do$/;"	module
+macro	input.ex	/^  defmacro macro(expr, opts) do$/;"	macro
+macro_no_params	input.ex	/^  defmacro macro_no_params do$/;"	macro
+private_macro	input.ex	/^  defmacrop private_macro(expr, opts) do$/;"	macro
+private_macro_no_params	input.ex	/^  defmacrop private_macro_no_params do$/;"	macro
