+++ Units/parser-elixir.r/elixir-macros.d/input.ex
@@ -1,0 +1,17 @@
+defmodule MacroModule do
+  defmacro macro(expr, opts) do
+    #
+  end
+
+  defmacro macro_no_params do
+    #
+  end
+
+  defmacrop private_macro(expr, opts) do
+    #
+  end
+
+  defmacrop private_macro_no_params do
+    #
+  end
+end
