+++ Units/parser-elixir.r/elixir-modules.d/expected.tags
@@ -1,0 +1,3 @@
+Foo	input.ex	/^defmodule Foo do$/;"	module
+Bar	input.ex	/^defmodule Foo.Bar do$/;"	module
+Baz	input.ex	/^  defmodule Baz do$/;"	module
