+++ Units/parser-elixir.r/elixir-modules.d/input.ex
@@ -1,0 +1,7 @@
+defmodule Foo do
+end
+
+defmodule Foo.Bar do
+  defmodule Baz do
+  end
+end
