+++ Units/parser-elixir.r/elixir-records.d/expected.tags
@@ -1,0 +1,3 @@
+RecordsModule	input.ex	/^defmodule RecordsModule do$/;"	module
+user1	input.ex	/^  Record.defrecord(:user1, name: "megan", age: "25")$/;"	record
+user2	input.ex	/^  Record.defrecordp :user2, name: "ivan", age: "23"$/;"	record
