+++ Units/parser-elixir.r/elixir-sign-operators.d/expected.tags
@@ -1,0 +1,20 @@
+OperatorModule	input.ex	/^defmodule OperatorModule do$/;"	module
++	input.ex	/^  def a + b, do: max(a, b)$/;"	operator
+-	input.ex	/^  def a - b, do: max(a, b)$/;"	operator
+*	input.ex	/^  def a * b, do: max(a, b)$/;"	operator
+/	input.ex	/^  def a \/ b, do: max(a, b)$/;"	operator
+=	input.ex	/^  def a = b, do: max(a, b)$/;"	operator
+.	input.ex	/^  def a . b, do: max(a, b)$/;"	operator
+|	input.ex	/^  def a | b, do: max(a, b)$/;"	operator
+|||	input.ex	/^  def _ ||| b, do: max(a, b)$/;"	operator
+&&&	input.ex	/^  def a &&& _, do: max(a, b)$/;"	operator
+<<<	input.ex	/^  def a <<< b, do: max(a, b)$/;"	operator
+>>>	input.ex	/^  defp a >>> b, do: max(a, b)$/;"	operator
+<<~	input.ex	/^  defp _ <<~ _, do: max(a, b)$/;"	operator
+~>>	input.ex	/^  defp a ~>> b, do: max(a, b)$/;"	operator
+<~	input.ex	/^  defmacro a <~ b, do: max(a, b)$/;"	operator
+~>	input.ex	/^  defmacro _ ~> b, do: max(a, b)$/;"	operator
+<~>	input.ex	/^  defmacro a <~> _, do: max(a, b)$/;"	operator
+<|>	input.ex	/^  defmacrop a <|> b, do: max(a, b)$/;"	operator
+^^^	input.ex	/^  defmacrop _ ^^^ _, do: max(a, b)$/;"	operator
+~~~	input.ex	/^  defmacrop a ~~~ b, do: max(a, b)$/;"	operator
