+++ Units/parser-elixir.r/elixir-tests.d/expected.tags
@@ -1,0 +1,3 @@
+TestModule	input.ex	/^defmodule TestModule do$/;"	module
+good with spaces	input.ex	/^  test "good with spaces" do$/;"	test
+good with parens	input.ex	/^  test("good with parens") do$/;"	test
